Replace method with a class in the command permission checker
Instead of using a method to check
if a user is allowed to use a command,
the programmer now should provide
a class which
implement IPermissionChecker.

In the class, the property message
is used to show a message to the user when
he is not allowed to use the command.

Fix #123